% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-02_s4_nodes_and_s4_edges.R
\docType{class}
\name{ancestor.edges-class}
\alias{ancestor.edges-class}
\alias{ancestor.edges}
\title{ancestor.edges S4 class}
\description{
ancestor.edges S4 class
}
\section{Slots}{

\describe{
\item{\code{data}}{edges dataframe}

\item{\code{required_fields}}{required fields for a edges s4 class object}

\item{\code{attribute_fields}}{fields denoting edge attributes}
}}

