% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-03_nodes.and.edges.R
\docType{class}
\name{ancestor.nodes.and.edges-class}
\alias{ancestor.nodes.and.edges-class}
\alias{ancestor.nodes.and.edges}
\title{ancestor.nodes.and.edges Reference Class}
\description{
ancestor.nodes.and.edges Reference Class
}

