% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_concept_examples.R
\name{append_concept_examples}
\alias{append_concept_examples}
\title{Append omop.graph with Concepts}
\usage{
append_concept_examples(
  omop_graph,
  sample_size = 5,
  schema = "omop_vocabulary",
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  verbose = FALSE,
  render_sql = FALSE
)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}

\item{sample_size}{PARAM_DESCRIPTION, Default: 5}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Example concepts are added to see
individuals belonging to a given class. This operation
can only be performed once and if it has already been appended
before, an error will be returned.
}
\details{
DETAILS
}
