% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chariotViz.R
\name{chariotViz}
\alias{chariotViz}
\title{FUNCTION_TITLE}
\usage{
chariotViz(
  omop_graph,
  force = FALSE,
  layout = NULL,
  output = NULL,
  as_svg = FALSE,
  title = NULL,
  width = 1000,
  height = 1000
)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}

\item{layout}{PARAM_DESCRIPTION, Default: NULL}

\item{output}{PARAM_DESCRIPTION, Default: NULL}

\item{as_svg}{PARAM_DESCRIPTION, Default: FALSE}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{width}{PARAM_DESCRIPTION, Default: NULL}

\item{height}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DiagrammeR]{render_graph}}
}
