% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{complete.nodes-class}
\alias{complete.nodes-class}
\alias{complete.nodes}
\title{complete.nodes S4 class}
\description{
complete.nodes S4 class
}
\section{Slots}{

\describe{
\item{\code{data}}{nodes dataframe}

\item{\code{required_fields}}{required fields for a nodes s4 class object}

\item{\code{attribute_fields}}{fields denoting node attributes}
}}

