% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-01_relationships_and_ancestors.R
\docType{class}
\name{complete.omop.relationships-class}
\alias{complete.omop.relationships-class}
\alias{complete.omop.relationships}
\title{complete.omop.relationships Reference Class}
\description{
complete.omop.relationships Reference Class
}

