% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_edge_legend.R
\name{construct_edge_legend}
\alias{construct_edge_legend}
\title{FUNCTION_TITLE}
\usage{
construct_edge_legend(omop_graph)
}
\arguments{
\item{graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
