% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{construct_graph}
\alias{construct_graph}
\title{FUNCTION_TITLE}
\usage{
construct_graph(nodes_and_edges, attr_theme = "lr")
}
\arguments{
\item{nodes_and_edges}{PARAM_DESCRIPTION}

\item{attr_theme}{PARAM_DESCRIPTION, Default: 'lr'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
