% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_complete_nodes_and_edges.R
\name{create_complete_nodes_and_edges}
\alias{create_complete_nodes_and_edges}
\title{FUNCTION_TITLE}
\usage{
create_complete_nodes_and_edges(
  complete_omop_relationships,
  type_from = invalid_reason,
 
    label_glue = "{ifelse(is.na(invalid_reason), 'V', invalid_reason)}\\n{vocabulary_id}\\n{concept_class_id}\\n({standard_concept})\\n"
)
}
\arguments{
\item{type_from}{PARAM_DESCRIPTION, Default: concept_class_id}

\item{omop_relationships}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
