% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{create_legend}
\alias{create_legend}
\title{FUNCTION_TITLE}
\usage{
create_legend(graph)
}
\arguments{
\item{graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
