% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{create_nodes_and_edges}
\alias{create_nodes_and_edges}
\title{FUNCTION_TITLE}
\usage{
create_nodes_and_edges(
  omop_relationships,
  type_from = concept_class_id,
  label_glue = "{vocabulary_id}\\n{concept_class_id}\\n({standard_concept})\\n"
)
}
\arguments{
\item{omop_relationships}{PARAM_DESCRIPTION}

\item{type_from}{PARAM_DESCRIPTION, Default: concept_class_id}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
