% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_complete_omop_relationships}
\alias{fetch_complete_omop_relationships}
\title{Fetch Relationships with Deprecated Concepts}
\usage{
fetch_complete_omop_relationships(
  ...,
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  schema = "omop_vocabulary",
  verbose = FALSE,
  render_sql = FALSE,
  version_key
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{version_key}{A list object that serves
as the hash for caching. Set to NULL to cache under a generic hash.}

\item{type_from}{PARAM_DESCRIPTION, Default: concept_class_id}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The difference between
this function and \code{fetch_omop_relationships} is that all
concepts belonging to a vocabulary are retrieved, including
deprecated concepts. In \code{fetch_omop_relationships}, \code{concept_class_id} matches
on both sides are filtered out, but here, a deprecated concept may map to a valid
concept within the concept concept class. For this reason, the filtering is adjusted
by removing relationships where concept_ids are equal instead. Invalid relationships
in the Concept Relationship table continue to be filtered out.

A new field \code{validity_status} is introduced here, which is \code{invalid_reason} with
the NULL value recorded to 'V'.
}
\details{
The OMOP vocabularies are transformed into nodes
and edges by first normalizing
OMOP concepts to their associated concept classes.
The distinct \code{concept_id} count
is preserved to determine the extent of coverage
between concept classes across
relationships.

This function queries excludes invalid concepts
and concept relationships as well
as relationships to self.
}
