% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{fetch_heatmap}
\alias{fetch_heatmap}
\title{FUNCTION_TITLE}
\usage{
fetch_heatmap(
  vocabulary_id_1,
  schema = "omop_vocabulary",
  verbose = FALSE,
  render_sql = FALSE,
  conn_fun = "pg13::local_connect(verbose=FALSE)"
)
}
\arguments{
\item{vocabulary_id_1}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}
}
\description{
FUNCTION_DESCRIPTION
}
