% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_omop_ancestors}
\alias{fetch_omop_ancestors}
\title{Fetch OMOP Ancestors}
\usage{
fetch_omop_ancestors(
  ...,
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  type_from = concept_class_id,
  label_glue = "{vocabulary_id}\\n{concept_class_id}\\n({standard_concept})\\n",
  schema = "omop_vocabulary",
  verbose = FALSE,
  render_sql = FALSE,
  version_key
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{type_from}{PARAM_DESCRIPTION, Default: concept_class_id}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{version_key}{A list object that serves
as the hash for caching. Set to NULL to cache under a generic hash.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Fetch concept class ancestry in a
Postgres instance of the OMOP vocabularies.
}
