% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{filter_graph}
\alias{filter_graph}
\title{FUNCTION_TITLE}
\usage{
filter_graph(omop_graph, ...)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{filter}}
}
