% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_node_id_1}
\alias{filter_node_id_1}
\title{FUNCTION_TITLE}
\usage{
filter_node_id_1(omop_graph, ...)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
