% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ancestors.R
\name{filter_omop_ancestors}
\alias{filter_omop_ancestors}
\title{FUNCTION_TITLE}
\usage{
filter_omop_ancestors(omop_ancestors, ...)
}
\arguments{
\item{omop_ancestors}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
