% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_omop_relationships.R
\name{filter_out_lateral_relationships}
\alias{filter_out_lateral_relationships}
\title{FUNCTION_TITLE}
\usage{
filter_out_lateral_relationships(omop_relationships)
}
\arguments{
\item{omop_relationships}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
