% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_key.R
\name{format_version_key}
\alias{format_version_key}
\title{FUNCTION_TITLE}
\usage{
format_version_key(version_key, ...)
}
\arguments{
\item{version_key}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
