% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_key.R
\name{get_vocabulary_version_key}
\alias{get_vocabulary_version_key}
\title{FUNCTION_TITLE}
\usage{
get_vocabulary_version_key(
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  schema = "omop_vocabulary",
  template_only = FALSE,
  sql_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{template_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{sql_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
\code{\link[pg13]{query}}
}
\keyword{internal}
