% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{include_legend}
\alias{include_legend}
\title{FUNCTION_TITLE}
\usage{
include_legend(omop_graph)
}
\arguments{
\item{graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
