% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_edges}
\alias{mutate_edges}
\title{FUNCTION_TITLE}
\usage{
mutate_edges(nodes_and_edges, ...)
}
\arguments{
\item{nodes_and_edges}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
