% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-01_relationships_and_ancestors.R
\docType{class}
\name{omop.ancestors-class}
\alias{omop.ancestors-class}
\alias{omop.ancestors}
\title{omop.ancestors Reference class}
\description{
omop.ancestors Reference class
}

