% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-03_nodes.and.edges.R, R/dev-classes.R
\docType{class}
\name{omop.concept.example-class}
\alias{omop.concept.example-class}
\alias{omop.concept.example}
\title{omop.concept.example S4 class}
\description{
omop.concept.example S4 class

omop.concept.example S4 class
}
