% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-01_relationships_and_ancestors.R
\docType{class}
\name{omop.relationships-class}
\alias{omop.relationships-class}
\alias{omop.relationships}
\title{omop.relationships Reference class}
\description{
omop.relationships Reference class
}

