% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{FUNCTION_TITLE}
\usage{
plot_heatmap(
  data,
  x = label_2,
  y = relationship_id,
  fill = label_1_coverage,
  fontsize = 6
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION, Default: label_2}

\item{y}{PARAM_DESCRIPTION, Default: relationship_id}

\item{fill}{PARAM_DESCRIPTION, Default: label_1_coverage}

\item{fontsize}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_raster}},\code{\link[ggplot2]{scale_colour_viridis_d}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
}
