% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_omop_relationships}
\alias{print_omop_relationships}
\title{FUNCTION_TITLE}
\usage{
print_omop_relationships(x, max.rows = 3)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
