% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap_fillcolor_by_concept_class}
\alias{remap_fillcolor_by_concept_class}
\title{FUNCTION_TITLE}
\usage{
remap_fillcolor_by_concept_class(
  omop_graph,
  fillcolor_from = concept_class_id,
  fillcolor_map = node_color_map@concept_class_id \%>\% purrr::set_names(NULL) \%>\%
    unlist(),
  fillcolor_map_other = "gray20"
)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
