% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_fillcolor_by_domain.R
\name{remap_fillcolor_by_domain}
\alias{remap_fillcolor_by_domain}
\title{FUNCTION_TITLE}
\usage{
remap_fillcolor_by_domain(omop_graph)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
