% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_key.R
\name{rmd_print_version}
\alias{rmd_print_version}
\title{FUNCTION_TITLE}
\usage{
rmd_print_version(version_key, ...)
}
\arguments{
\item{version_key}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
