% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_chariotViz}
\alias{setup_chariotViz}
\title{FUNCTION_TITLE}
\usage{
setup_chariotViz(
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  schema = "omop_vocabulary",
  verbose = FALSE,
  render_sql = FALSE,
  version_key
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{version_key}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
