% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{view_to_counts}
\alias{view_to_counts}
\title{FUNCTION_TITLE}
\usage{
view_to_counts(omop_graph, ...)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{context}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{tidyeval-compat}}
}
