% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-02_s4_nodes_and_s4_edges.R
\docType{class}
\name{ancestor.nodes-class}
\alias{ancestor.nodes-class}
\alias{ancestor.nodes}
\title{ancestor.nodes S4 class}
\description{
ancestor.nodes S4 class
}
\section{Slots}{

\describe{
\item{\code{data}}{nodes dataframe}

\item{\code{required_fields}}{required fields for a nodes s4 class object}

\item{\code{attribute_fields}}{fields denoting node attributes}
}}

