% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{complete.omop.relationships-class}
\alias{complete.omop.relationships-class}
\alias{complete.omop.relationships}
\title{complete.omop.relationships S4 class}
\description{
complete.omop.relationships S4 class
}
