% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap_fillcolor_by_domain}
\alias{remap_fillcolor_by_domain}
\title{FUNCTION_TITLE}
\usage{
remap_fillcolor_by_domain(
  omop_graph,
  fillcolor_from = domain_id,
  fillcolor_map = node_color_map@domain_id$Base,
  fillcolor_map_other = "gray20"
)
}
\arguments{
\item{omop_graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
